refactor: move some functions to setup
